Move routes to default.py, Add API_BASE to BASE_URL, Rename logout function
