+++ resources/lib/WatchlistFlavor/WonderfulSubs.py
@@ -6,8 +6,7 @@ from WatchlistFlavorBase import WatchlistFlavorBase
 
 
 class WonderfulSubsWLF(WatchlistFlavorBase):
-    _URL = BASE_URL
-    _API_BASE = API_BASE
+    _URL = "{}/{}".format(BASE_URL, API_BASE)
     _TITLE = "WonderfulSubs"
     _NAME = "wonderfulsubs"
 
@@ -16,7 +15,7 @@ class WonderfulSubsWLF(WatchlistFlavorBase):
         return self._login_image
 
     def login(self):
-        url = self._to_url("{}/users/login".format(self._API_BASE))
+        url = self._to_url("users/login")
         data = (self._post_request(url, json={"username": self._username, "password": self._password})).json()
 
         if data['success'] is not True:
@@ -27,10 +26,7 @@ class WonderfulSubsWLF(WatchlistFlavorBase):
                                        ('%s/%s' % (data['data']['_id'], data['token'])))
 
     def watchlist(self):
-        url = self._to_url("%s/watchlist/list?_id=%s" % (
-            self._API_BASE,
-            self._login_token.split("/")[0]
-        ))
+        url = self._to_url("watchlist/list?_id=%s" % self._login_token.split("/")[0])
         return self._process_watchlist_view(url, "watchlist/%d", page=1)
 
     def __header(self):
