+++ resources/lib/WonderfulSubsBrowser.py
@@ -11,7 +11,7 @@ from .login import login
 
 
 class WonderfulSubsBrowser(BrowserBase):
-    _BASE_URL = BASE_URL
+    _BASE_URL = "{}/{}".format(BASE_URL, API_BASE)
     _RESULTS_PER_SEARCH_PAGE = 25
 
     def __init__(self, base_flavor):
@@ -191,14 +191,14 @@ class WonderfulSubsBrowser(BrowserBase):
             "platform": "Kodi",
         }
         link = "%s?%s" % (
-            self._to_url("{}/media/stream".format(API_BASE)),
+            self._to_url("media/stream"),
             urllib.urlencode(video_data)
         )
         return {sname: link}
 
     def _get_anime_info_obj(self, anime_url):
         results = self._json_request(
-            self._to_url("{}/media/series".format(API_BASE)),
+            self._to_url("media/series"),
             {"series": anime_url}
         )
 
@@ -294,7 +294,7 @@ class WonderfulSubsBrowser(BrowserBase):
             "index": (page-1) * self._RESULTS_PER_SEARCH_PAGE,
         }
 
-        url = self._to_url("{}/media/search".format(API_BASE))
+        url = self._to_url("media/search")
         return self._process_anime_view(url, data, "search/%s/%%d" % search_string, page)
 
     # TODO: Not sure i want this here..
@@ -310,7 +310,7 @@ class WonderfulSubsBrowser(BrowserBase):
             "count": self._RESULTS_PER_SEARCH_PAGE,
             "index": (page-1) * self._RESULTS_PER_SEARCH_PAGE,
         }
-        url = self._to_url("{}/media/all".format(API_BASE))
+        url = self._to_url("media/all")
         return self._process_anime_view(url, data, "letter/%s/%%d" % letter, page)
 
     def get_all(self,  page=1):
@@ -318,7 +318,7 @@ class WonderfulSubsBrowser(BrowserBase):
             "count": self._RESULTS_PER_SEARCH_PAGE,
             "index": (page-1) * self._RESULTS_PER_SEARCH_PAGE,
         }
-        url = self._to_url("{}/media/all".format(API_BASE))
+        url = self._to_url("media/all")
         return self._process_anime_view(url, data, "all/%d", page)
 
     def get_popular(self,  page=1):
@@ -326,7 +326,7 @@ class WonderfulSubsBrowser(BrowserBase):
             "count": self._RESULTS_PER_SEARCH_PAGE,
             "index": (page-1) * self._RESULTS_PER_SEARCH_PAGE,
         }
-        url = self._to_url("{}/media/popular".format(API_BASE))
+        url = self._to_url("media/popular")
         return self._process_anime_view(url, data, "popular/%d", page)
 
     def get_latest(self, page=1):
@@ -334,7 +334,7 @@ class WonderfulSubsBrowser(BrowserBase):
             "count": self._RESULTS_PER_SEARCH_PAGE,
             "index": (page-1) * self._RESULTS_PER_SEARCH_PAGE,
         }
-        url = self._to_url("{}/media/latest".format(API_BASE))
+        url = self._to_url("media/latest")
         return self._process_anime_view(url, data, "latest/%d", page)
 
     def get_random(self, page=1):
@@ -342,7 +342,7 @@ class WonderfulSubsBrowser(BrowserBase):
             "count": self._RESULTS_PER_SEARCH_PAGE,
             "index": (page-1) * self._RESULTS_PER_SEARCH_PAGE,
         }
-        url = self._to_url("{}/media/random".format(API_BASE))
+        url = self._to_url("media/random")
         return self._process_anime_view(url, data, "random/%d", page)
 
     def get_anime_metadata(self, anime_url, is_dubbed):
